P0 工程基线：vendor 入库、CI 执法链、README 一致性、CHANGELOG
- 修复 fresh clone 不可构建（阻断级）：vendor/{pi-shell,pi-builtins,pi-walker,brush-core}
  入库（.gitignore 改内容级排除），删除 982MB 无引用 cargo-vendor 残留
- CI 三道执法门：新增 lint（cargo fmt --check + clippy -D warnings）与 test 门；
  移除 RUSTFLAGS=-A warnings；release 增加产物 --version 冒烟与 SHA256 校验和 sidecar
- lint 基线校准：pedantic/nursery 整组降为 allow（83k 行存量从未在其下维护，噪音 600+），
  默认组全执法 + correctness/perf deny + 6 个高信号 pedantic 项显式入选；
  移除 27 处 lib.rs 属性 lint 墙（属性级别高于 Cargo.toml，会架空整份 allow 名单）；
  clippy 存量 1775 → 0
- README×3 一致性：修 MCP 能力表述（stdio 传输）、工具注册机制表述、模块树 23→33 crates；
  clone/repo URL 落定 brianmiller123/Gyre；修复 tests/understand/README.md 断链
- 新增 CHANGELOG.md、.config/nextest.toml（retries=0）、rustfmt.toml；docs/ 解除忽略
- 附：08-06 以来的全部功能 WIP 首次入库（进程内 shell、Gemini 适配器、hub、
  security_scan、todo/ask/checkpoint 工具、记忆升级、429 Retry-After 等）

+++ crates/hashline/src/format.rs
@@ -85,7 +85,7 @@ pub fn format_numbered_lines(text: &str, start_line: Anchor) -> String {
 pub fn compute_file_hash(text: &str) -> String {
     let normalized: String = normalize_for_hash(text);
     let h = fnv1a_32(normalized.as_bytes());
-    let low16 = (h & 0xFFFF) as u32;
+    let low16 = h & 0xFFFF;
     format!("{low16:0HASH_LEN$X}")
 }
